+++ app/builder.py
@@ -4,12 +4,20 @@ from app.site.page_agencies import AgenciesPage
 from app.site.page_headlines import HeadlinesPage
 from app.site.page_topics import TopicsPage
 from app.utils.logger import get_logger
+from app.site.data import DataHandler
+from app.site.graphing import Plots
 
 logger = get_logger(__name__)
 
 
+def gen_plots(dh: DataHandler):
+    Plots.sentiment_graphs(dh.all_sentiment_data)
+
+
 def build():
+    dh: DataHandler = DataHandler()
     clear_build()
+    gen_plots(dh)
     pages = [HeadlinesPage, AgenciesPage, TopicsPage]
     for page in pages:
         page().generate()
