+++ tests/conftest.py
@@ -1,7 +1,9 @@
 import os
+import pytest
 
 from app.utils.config import Config
 from app.utils.constants import Constants
+from app.site.data import DataHandler
 
 
 def pytest_addoption(parser):
@@ -13,3 +15,8 @@ def pytest_sessionstart(session):
         os.rename(os.path.join(Constants.Paths.ROOT, 'tests', 'test.db'), Config.db_file_path)
     Config.set_debug()
     assert Config.debug
+
+
+@pytest.fixture(scope='session')
+def data_handler():
+    return DataHandler()
